Refac add buttons
